refactor: rename commonMain/login to commonMain/auth
