+++ androidApp/src/main/java/com/example/tasky/android/agenda/viewmodel/AgendaViewModel.kt
@@ -13,10 +13,10 @@ import com.example.tasky.android.agenda.screen.AgendaDetailsScreenType
 import com.example.tasky.android.agenda.screen.AgendaItemUi
 import com.example.tasky.android.agenda.screen.AgendaScreenEvent
 import com.example.tasky.android.agenda.screen.AgendaScreenState
+import com.example.tasky.auth.domain.ILoginRepository
+import com.example.tasky.auth.domain.manager.SessionManager
+import com.example.tasky.auth.domain.util.getAvatarDisplayName
 import com.example.tasky.common.domain.model.onSuccess
-import com.example.tasky.login.domain.ILoginRepository
-import com.example.tasky.login.domain.manager.SessionManager
-import com.example.tasky.login.domain.util.getAvatarDisplayName
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
