+++ androidApp/src/main/java/com/example/tasky/android/login/viewmodel/RegisterViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasky.android.login.screen.RegisterScreenEvent
 import com.example.tasky.android.login.screen.RegisterScreenState
+import com.example.tasky.auth.domain.ILoginRepository
+import com.example.tasky.auth.domain.util.Validator
 import com.example.tasky.common.domain.model.onError
 import com.example.tasky.common.domain.model.onSuccess
-import com.example.tasky.login.domain.ILoginRepository
-import com.example.tasky.login.domain.util.Validator
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
