+++ shared/src/commonMain/kotlin/com/example/tasky/auth/data/LoginRepository.kt
@@ -1,5 +1,8 @@
-package com.example.tasky.login.data
+package com.example.tasky.auth.data
 
+import com.example.tasky.auth.data.model.LoginBody
+import com.example.tasky.auth.data.model.RegisterBody
+import com.example.tasky.auth.domain.ILoginRepository
 import com.example.tasky.common.data.manager.HttpManager
 import com.example.tasky.common.data.model.BaseError
 import com.example.tasky.common.domain.model.ResultWrapper
@@ -8,9 +11,6 @@ import com.example.tasky.common.domain.model.onSuccess
 import com.example.tasky.dataStore.SettingsKey
 import com.example.tasky.dataStore.createSettings
 import com.example.tasky.dataStore.dataStore
-import com.example.tasky.login.data.model.LoginBody
-import com.example.tasky.login.data.model.RegisterBody
-import com.example.tasky.login.domain.ILoginRepository
 import com.russhwolf.settings.ExperimentalSettingsApi
 import com.russhwolf.settings.ExperimentalSettingsImplementation
 import com.russhwolf.settings.coroutines.FlowSettings
