+++ shared/src/commonMain/kotlin/com/example/tasky/auth/domain/util/Validator.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.login.domain.util
+package com.example.tasky.auth.domain.util
 
 object Validator {
     private val emailAddressRegex =
