+++ shared/src/commonMain/kotlin/com/example/tasky/common/data/manager/HttpManager.kt
@@ -1,10 +1,10 @@
 package com.example.tasky.common.data.manager
 
 import com.example.tasky.BuildKonfig.API_KEY
+import com.example.tasky.auth.data.model.AccessTokenResponse
+import com.example.tasky.auth.domain.manager.SessionManager
 import com.example.tasky.common.data.model.ErrorResponse
 import com.example.tasky.common.data.util.isSuccess
-import com.example.tasky.login.data.model.AccessTokenResponse
-import com.example.tasky.login.domain.manager.SessionManager
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.plugins.HttpTimeout
@@ -18,7 +18,6 @@ import io.ktor.client.plugins.logging.Logger
 import io.ktor.client.plugins.logging.Logging
 import io.ktor.client.plugins.resources.post
 import io.ktor.client.request.header
-import io.ktor.client.request.host
 import io.ktor.client.request.setBody
 import io.ktor.http.ContentType
 import io.ktor.http.URLProtocol
@@ -72,7 +71,6 @@ internal object HttpManager {
 
                             if (!response.isSuccess()) {
                                 val errorResponse = response.body<ErrorResponse>()
-                                println("refresh token fail: ${errorResponse.message}")
                                 throw Exception(errorResponse.message)
                             }
 
@@ -83,6 +81,8 @@ internal object HttpManager {
                             BearerTokens(responseBody.accessToken, body.refreshToken)
                         } catch (e: Exception) {
                             if (e is CancellationException) throw e
+
+                            println("refresh token fail: ${e.message}")
 
                             SessionManager.removeToken()
                             Throwable(e)
